`transport` files, mods renaming, remove schema cursor setting
