+++ bin/miden-cli/src/commands/mod.rs
@@ -5,8 +5,8 @@ pub mod import;
 pub mod init;
 pub mod new_account;
 pub mod new_transactions;
+pub mod note_transport;
 pub mod notes;
 pub mod sync;
 pub mod tags;
 pub mod transactions;
-pub mod transport;
