+++ bin/miden-cli/src/commands/note_transport.rs
