+++ crates/idxdb-store/src/js/schema.js
