--- crates/idxdb-store/src/js/transport.d.ts
@@ -1,2 +1,0 @@
-export declare function getNoteTransportCursor(): Promise<number>;
-export declare function updateNoteTransportCursor(cursor: number): Promise<void>;
