+++ crates/rust-client/src/note_transport/errors.rs
