+++ crates/rust-client/src/note_transport/grpc.rs
