+++ crates/rust-client/src/test_utils/mod.rs
@@ -1,5 +1,5 @@
 pub mod mock;
-pub mod transport;
+pub mod note_transport;
 
 #[cfg(feature = "std")]
 pub mod common;
